Whats_app_AI_Bot/requirements.txt
Flask
twilio
openai
google-cloud-translate
pydub
faster-whisper
langdetect
ollama
python-dotenv